requirements.txt
###Core dependencies
psutil>=5.9.0        # System monitoring
PyYAML>=6.0.1       # Configuration file handling
rich>=13.3.1        # Console output formatting

###Optional dependencies
pandas>=2.0.0       # For future data analysis features
matplotlib>=3.7.1   # For future graphing features
